Add files via upload
This is my undergraduate research paper completed in the year 2021 under Prof.Anand Prakash. I looked at the case of Anaemia prevalence among women in India with the help of all the rounds of NFHS data. The goal of the research was to understand the reasons behind incidence of anaemia among women belonging to different socio-economic groups in India by looking at the food consumption pattern of the women.