Add files via upload
This is a policy intervention project to understand the "Policy Toolkit" given by Kelkar and Shah in their book "In Service of the republic. I took the case of Anaemia among children in Jharkhand and designed a policy through the mid-day meal scheme to understand if to improve improve on their current loopholes.