Add files via upload
Published this article while doing a research internship with the organisation Foundation to Aid Industrial Recovery" under the guidance of Prof.Sashi Sivramkrishna. During the period of internship, it helped me understand the concept of modern money theory and how it applies in the case of austerity of an economy.